+++ origin/src/Console/Shell.php
@@ -18,6 +18,7 @@ use Origin\Console\ConsoleInput;
 use Origin\Console\ConsoleOutput;
 use Origin\Console\Task\TaskRegistry;
 use Origin\Model\ModelRegistry;
+use Origin\Core\Logger;
 
 use ReflectionClass;
 use ReflectionMethod;
@@ -285,5 +286,16 @@ class Shell
     public function taskRegistry()
     {
         return $this->taskRegistry;
+    }
+
+    /**
+     * Returns a Logger Object
+     *
+     * @param string $channel
+     * @return Logger
+     */
+    public function logger(string $channel = 'Shell')
+    {
+        return new Logger($channel);
     }
 }