+++ origin/src/Console/Task/Task.php
@@ -16,6 +16,7 @@ namespace Origin\Console\Task;
 
 use Origin\Console\Shell;
 use Origin\Core\ConfigTrait;
+use Origin\Core\Logger;
 
 class Task
 {
@@ -147,5 +148,16 @@ class Task
     public function out(string $data, $newLine = true)
     {
         $this->shell()->out($data, $newLine);
+    }
+
+    /**
+     * Returns a Logger Object
+     *
+     * @param string $channel
+     * @return Logger
+     */
+    public function logger(string $channel = 'Task')
+    {
+        return new Logger($channel);
     }
 }