+++ origin/src/Model/Model.php
@@ -19,6 +19,7 @@ use Origin\Model\Behavior\BehaviorRegistry;
 use Origin\Model\Exception\MissingModelException;
 use Origin\Exception\NotFoundException;
 use Origin\Exception\InvalidArgumentException;
+use Origin\Core\Logger;
 
 class Model
 {
@@ -1778,5 +1779,16 @@ class Model
         }
 
         return true;
+    }
+
+    /**
+     * Returns a Logger Object
+     *
+     * @param string $channel
+     * @return Logger
+     */
+    public function logger(string $channel = 'Model')
+    {
+        return new Logger($channel);
     }
 }